+++ plugin/mvn.go
@@ -1,0 +1,100 @@
+package plugin
+
+import (
+	"log"
+)
+
+var MavenRunCmdJsonTagToExeFlagMapStringItemList = []JsonTagToExeFlagMapStringItem{
+	{"--build-name=", "PLUGIN_BUILD_NAME", false, false},
+	{"--build-number=", "PLUGIN_BUILD_NUMBER", false, false},
+	{"--detailed-summary=", "PLUGIN_DETAILED_SUMMARY", false, false},
+	{"--format=", "PLUGIN_FORMAT", false, false},
+	{"--insecure-tls=", "PLUGIN_INSECURE", false, false},
+	{"--project=", "PLUGIN_PROJECT", false, false},
+	{"--scan=", "PLUGIN_SCAN", false, false},
+	{"--threads=", "PLUGIN_THREADS", false, false},
+}
+
+var MavenConfigCmdJsonTagToExeFlagMapStringItemList = []JsonTagToExeFlagMapStringItem{
+	{"--exclude-patterns=", "PLUGIN_EXCLUDE_PATTERNS", false, false},
+	{"--global=", "PLUGIN_GLOBAL", false, false},
+	{"--include-patterns=", "PLUGIN_INCLUDE_PATTERNS", false, false},
+	{"--repo-deploy-releases=", "PLUGIN_DEPLOY_RELEASE_REPO", false, false},
+	{"--repo-deploy-snapshots=", "PLUGIN_DEPLOY_SNAPSHOT_REPO", false, false},
+	{"--repo-resolve-releases=", "PLUGIN_RESOLVE_RELEASE_REPO", false, false},
+	{"--repo-resolve-snapshots=", "PLUGIN_RESOLVE_SNAPSHOT_REPO", false, false},
+	{"--server-id-deploy=", "PLUGIN_SERVER_ID_DEPLOY", false, false},
+	{"--server-id-resolve=", "PLUGIN_RESOLVER_ID", false, false},
+	{"--use-wrapper=", "PLUGIN_USE_WRAPPER", false, false},
+}
+
+func GetMavenBuildCommandArgs(args Args) ([][]string, error) {
+
+	var cmdList [][]string
+
+	jfrogConfigAddConfigCommandArgs, err := GetConfigAddConfigCommandArgs(args.ResolverId,
+		args.Username, args.Password, args.URL, args.AccessToken, args.APIKey)
+	if err != nil {
+		return cmdList, err
+	}
+
+	mvnConfigCommandArgs := []string{MvnConfig}
+	err = PopulateArgs(&mvnConfigCommandArgs, &args, MavenConfigCmdJsonTagToExeFlagMapStringItemList)
+	if err != nil {
+		return cmdList, err
+	}
+
+	mvnRunCommandArgs := []string{MvnCmd, args.MvnGoals}
+	err = PopulateArgs(&mvnRunCommandArgs, &args, MavenRunCmdJsonTagToExeFlagMapStringItemList)
+	if err != nil {
+		return cmdList, err
+	}
+	if len(args.MvnPomFile) > 0 {
+		mvnRunCommandArgs = append(mvnRunCommandArgs, "-f "+args.MvnPomFile)
+	}
+
+	cmdList = append(cmdList, jfrogConfigAddConfigCommandArgs)
+	cmdList = append(cmdList, mvnConfigCommandArgs)
+	cmdList = append(cmdList, mvnRunCommandArgs)
+
+	return cmdList, nil
+}
+
+var RtBuildInfoPublishCmdJsonTagToExeFlagMap = []JsonTagToExeFlagMapStringItem{
+	{"--project=", "PLUGIN_PROJECT", false, false},
+}
+
+func GetMavenPublishCommand(args Args) ([][]string, error) {
+
+	var cmdList [][]string
+	var jfrogConfigAddConfigCommandArgs []string
+
+	tmpServerId := args.DeployerId // "tmpSrvConfig"
+	jfrogConfigAddConfigCommandArgs, err := GetConfigAddConfigCommandArgs(tmpServerId,
+		args.Username, args.Password, args.URL, args.AccessToken, args.APIKey)
+	if err != nil {
+		log.Println("GetConfigAddConfigCommandArgs error: ", err)
+		return cmdList, err
+	}
+
+	mvnConfigCommandArgs := []string{MvnConfig}
+	err = PopulateArgs(&mvnConfigCommandArgs, &args, MavenConfigCmdJsonTagToExeFlagMapStringItemList)
+	if err != nil {
+		log.Println("PopulateArgs error: ", err)
+		return cmdList, err
+	}
+
+	rtPublishBuildInfoCommandArgs := []string{"rt", BuildPublish, args.BuildName, args.BuildNumber,
+		"--server-id=" + tmpServerId}
+	err = PopulateArgs(&rtPublishBuildInfoCommandArgs, &args, RtBuildInfoPublishCmdJsonTagToExeFlagMap)
+	if err != nil {
+		log.Println("PopulateArgs error: ", err)
+		return cmdList, err
+	}
+
+	cmdList = append(cmdList, jfrogConfigAddConfigCommandArgs)
+	cmdList = append(cmdList, mvnConfigCommandArgs)
+	cmdList = append(cmdList, rtPublishBuildInfoCommandArgs)
+
+	return cmdList, nil
+}
